Perferring to version the Graphql controller
