+++ app/controllers/api/v1x0/graphql_controller.rb
@@ -1,17 +1,6 @@
-require "manageiq/api/common/graphql"
-
 module Api
   module V1x0
-    class GraphqlController < ApplicationController
-      def query
-        graphql_api_schema = ::ManageIQ::API::Common::GraphQL::Generator.init_schema(request)
-        variables = ::ManageIQ::API::Common::GraphQL.ensure_hash(params[:variables])
-        result = graphql_api_schema.execute(
-          params[:query],
-          :variables => variables
-        )
-        render :json => result
-      end
+    class GraphqlController < Api::V1::GraphqlController
     end
   end
 end
