+++ public/doc/openapi-3-v1.0.0.json
@@ -1274,7 +1274,8 @@
                     },
                     "variables": {
                         "type": "object",
-                        "description": "Optional Query variables"
+                        "description": "Optional Query variables",
+                        "nullable": true
                     }
                 },
                 "required": [
